refactor(app): update request changes from review
* Separate skeleton loader
* Separate CafeMainImageCarousel
* Update variable name (cafeImagePanel -> cafeImagePanelDetail)
